+++ src/command/mod.rs
@@ -1,5 +1,7 @@
 pub mod argon;
 pub mod bootstrap;
+pub mod prune;
+pub mod replicate;
 pub mod scrub;
 pub mod server;
 pub mod worker;